Restructured files and updated packages
