+++ src/pages/layout.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { Outlet } from "react-router-dom";
+
+export default function RootLayout() {
+    return (
+        <div>
+            <Outlet />
+        </div>
+    );
+}
